requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pypdf
numpy
faiss-cpu
python-multipart
google-genai
rapidfuzz